made each app independent of themselves
