+++ project/settings.py
@@ -41,6 +41,7 @@ INSTALLED_APPS = [
     # PERSONAL APPS
     'playground.apps.PlaygroundConfig',
     'tags.apps.TagsConfig',
+    'store_custom.apps.StoreCustomConfig',
 
     # PIP INSTALLS
     'rest_framework',
